rtt_ros2_idl: automatically add RTT plugin dependencies in generate_typekit()
For the full functionality of a typekit generated for ROS interface packages also the typekits
for all nested types must be loaded at run-time. The ament resource index provides all the necessary
information about dependencies and which package provides interfaces.

All RTT plugin dependencies are also run-time dependencies (<exec_depend>) of the respective package,
although not declared in package.xml. The run-time dependencies are also exported as resource
'package_run_dependencies' in the ament index and used to order the packages topologically in the
generated setup scripts.